// src/app/models/utilisateur.ts
export interface Utilisateur {
    idUtilisateur: string
    prenom: string
    nom: string
    numero: string
    nonStructure: string
    addresse: string
    photo: string
    email: string
    refreshToken: string
    password: string
    roleUtilisateur: string
    dateModif: string
    actif: boolean
    dateInscription: string
}